# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/task-backend/tasks/user/private/tasks": {
      "post": {
        "operationId": "UserTasksController_createNewTask",
        "summary": "create a new task",
        "parameters": [
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Task - User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UserTasksController_paginateTasks",
        "summary": "paginate tasks",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "completed",
                "canceled"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedUser",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 20,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Task - User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/task-backend/tasks/user/private/tasks/{id}": {
      "get": {
        "operationId": "UserTasksController_getTaskByTaskId",
        "summary": "get task by taskId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Task - User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UserTasksController_updateSelfTask",
        "summary": "update task as manager",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Task - User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/task-backend/tasks/user/private/tasks/{id}/status": {
      "patch": {
        "operationId": "UserTasksController_updateTaskStatus",
        "summary": "update task status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Task - User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/task-backend/authentication/user/private-auth/refresh-token": {
      "post": {
        "operationId": "UserAuthController_refreshToken",
        "parameters": [
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "authentication-user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/task-backend/authentication/user/public/login-email": {
      "post": {
        "operationId": "UserAuthController_loginUser",
        "parameters": [
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "authentication-user"
        ]
      }
    },
    "/task-backend/authentication/user/public/forget-password": {
      "post": {
        "operationId": "UserAuthController_forgetPassword",
        "parameters": [
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "authentication-user"
        ]
      }
    },
    "/task-backend/authentication/user/public/verify-forget-password-email": {
      "post": {
        "operationId": "UserAuthController_verifyForgetPasswordEmail",
        "parameters": [
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "authentication-user"
        ]
      }
    },
    "/task-backend/authentication/user/reset-password": {
      "post": {
        "operationId": "UserAuthController_resetPassword",
        "parameters": [
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "authentication-user"
        ]
      }
    },
    "/task-backend/health": {
      "get": {
        "operationId": "HealthController_healthCheck",
        "parameters": [
          {
            "name": "x-api-version",
            "in": "header",
            "description": "API version",
            "required": true,
            "schema": {
              "type": "string",
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": {
                          "type": "string"
                        }
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": {
                          "type": "string"
                        }
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": {
                          "type": "string"
                        }
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": {
                          "type": "string"
                        }
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "health-check"
        ]
      }
    }
  },
  "info": {
    "title": "task-backend",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "issuedTo": {
            "example": [
              "689cf17b63e1c01b9f13d815"
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "dependencies": {
            "example": [
              "689cf98b53475ab90646d01a"
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "title": {
            "type": "string"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "issuedTo",
          "title",
          "dueDate"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "issuedTo": {
            "example": [
              "689cf17b63e1c01b9f13d815"
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "dependencies": {
            "example": [
              "689cf98b53475ab90646d01a"
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "title": {
            "type": "string"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UpdateTaskStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "canceled"
            ]
          }
        }
      },
      "LoginEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "rememberMe": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "pattern": "/^[A-Za-z0-9-_=]+\\.[A-Za-z0-9-_=]+\\.?[A-Za-z0-9-_.+/=]*$/"
          },
          "newPassword": {
            "type": "string",
            "pattern": "/^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#])[A-Za-z\\d@$!%*?&#]{8,}$/"
          }
        },
        "required": [
          "accessToken",
          "newPassword"
        ]
      }
    }
  }
}